refactor(vb-batch-17): split oversized files into module directories
This commit lands 4 long-standing file-split refactors that were in a
working state on the working tree but uncommitted:

1. crates/vb_core/src/budget.rs (was 2393 LoC) -> crates/vb_core/src/budget/
   directory (15 submodules: aggregate_budget, aggregate_usage,
   aggregate_usage_checks, budget_error, policy, small_linear, traversal,
   traversal_fanout, traversal_loop, traversal_path, traversal_step_count,
   traversal_successors, traversal_tracking, types, validation)

2. crates/vb_storage/src/tests.rs (was 8091 LoC) -> crates/vb_storage/src/tests/
   directory (41 chunk_001..chunk_041 modules; exact assertions preserved)

3. config + test infrastructure:
   - .config/source-length-exceptions.txt: remove obsolete budget/tests
     exception (now split)
   - .config/test-integrity-deletion-allow.txt: add chunk_001 redirection
     for the budget.rs -> budget/ split
   - .moon/tasks/all.yml: re-enable runInCI for check-test-determinism
     (now passes after baseline.txt was added)
   - scripts/check-test-determinism.py: add baseline comparison logic
     using .evidence/test-determinism/baseline.txt

4. New moon task definitions:
   - .moon/tasks/flux.yml: flux-check-vb-compile, flux-check-vb-runtime
   - .moon/tasks/loom.yml: loom-list-smoke, loom-run

5. Test file updates to follow the new include_str! targets:
   - crates/vb_core/benches/aggregate_resource_budget.rs: ../src/budget/mod.rs
   - crates/vb_core/tests/aggregate_resource_budget_red.rs: same
   - crates/vb_core/tests/aggregate_resource_budget_properties_red.rs: same
   - crates/vb_core/tests/aggregate_resource_budget_snapshot_red.rs: same
   - These tests include all 15 budget submodules so the source-string
     contract verification still covers the full symbol set.

Gates: cargo check 0 errors; cargo clippy 0 issues with -D warnings
-D unsafe_code; cargo test --workspace 14243 passed, 40 ignored.

Implements: vb-32gwc, vb-stortst01, vb-batch-17 (umbrella).

+++ crates/vb_core/src/kani_step_budget_zero.rs
@@ -76,7 +76,7 @@ fn kani_aggregate_usage_zero() {
     // up to and including the ceiling must satisfy the additivity property.
     let bounded_steps: u32 = kani::any();
     kani::assume(bounded_steps <= 1_000);
-    budget.max_steps_executable = u64::from(bounded_steps);
+    budget.max_steps_executable = bounded_steps;
     // After the assumption, the bounded value is in [0, MAX_STEPS_PER_WORKFLOW].
     // Adding to a zero usage cannot overflow because both operands are
     // bounded by the per-workflow ceiling.
